+++ engine/src/main/java/dev/gamekit/utils/Bounds.java
@@ -4,10 +4,10 @@ import java.lang.Math;
 
 /** {@link Bounds} represents a region in 2D space */
 public class Bounds {
-  public int x;
-  public int y;
-  public int width;
-  public int height;
+  public double x;
+  public double y;
+  public double width;
+  public double height;
 
   public Bounds() {
     this(0, 0, 0, 0);
@@ -18,7 +18,7 @@ public class Bounds {
     this(bounds.x, bounds.y, bounds.width, bounds.height);
   }
 
-  public Bounds(int x, int y, int width, int height) {
+  public Bounds(double x, double y, double width, double height) {
     this.x = x;
     this.y = y;
     this.width = width;
@@ -27,14 +27,14 @@ public class Bounds {
 
   /** Computes the intersection of two bounds and stores the result in {@code out} */
   public static void intersect(Bounds bounds1, Bounds bounds2, Bounds out) {
-    int x1 = java.lang.Math.max(bounds1.x, bounds2.x);
-    int y1 = java.lang.Math.max(bounds1.y, bounds2.y);
-    int x2 = java.lang.Math.min(bounds1.x + bounds1.width, bounds2.x + bounds2.width);
-    int y2 = java.lang.Math.min(bounds1.y + bounds1.height, bounds2.y + bounds2.height);
+    double x1 = java.lang.Math.max(bounds1.x, bounds2.x);
+    double y1 = java.lang.Math.max(bounds1.y, bounds2.y);
+    double x2 = java.lang.Math.min(bounds1.x + bounds1.width, bounds2.x + bounds2.width);
+    double y2 = java.lang.Math.min(bounds1.y + bounds1.height, bounds2.y + bounds2.height);
     out.set(x1, y1, x2 - x1, y2 - y1);
   }
 
-  public void set(int x, int y, int width, int height) {
+  public void set(double x, double y, double width, double height) {
     this.x = x;
     this.y = y;
     this.width = width;
@@ -42,24 +42,24 @@ public class Bounds {
   }
 
   /** Sets the position (x, y) of the bounds */
-  public void setPosition(int x, int y) {
+  public void setPosition(double x, double y) {
     this.x = x;
     this.y = y;
   }
 
   /** Sets the size (width, height) of the bounds */
-  public void setSize(int width, int height) {
+  public void setSize(double width, double height) {
     this.width = width;
     this.height = height;
   }
 
-  public void setX(int x) { this.x = x; }
+  public void setX(double x) { this.x = x; }
 
-  public void setY(int y) { this.y = y; }
+  public void setY(double y) { this.y = y; }
 
-  public void setWidth(int width) { this.width = width; }
+  public void setWidth(double width) { this.width = width; }
 
-  public void setHeight(int height) { this.height = height; }
+  public void setHeight(double height) { this.height = height; }
 
   /** Copies values from another bounds object */
   public void set(Bounds bounds) {
@@ -74,10 +74,10 @@ public class Bounds {
 
   /** Adjust this bounds to be a union of itself and {@code bounds} */
   public void extend(Bounds bounds) {
-    int x1 = java.lang.Math.min(x, bounds.x);
-    int y1 = java.lang.Math.min(y, bounds.y);
-    int w = java.lang.Math.max(x + width, bounds.x + bounds.width) - x1;
-    int h = Math.max(y + height, bounds.y + bounds.height) - y1;
+    double x1 = java.lang.Math.min(x, bounds.x);
+    double y1 = java.lang.Math.min(y, bounds.y);
+    double w = java.lang.Math.max(x + width, bounds.x + bounds.width) - x1;
+    double h = Math.max(y + height, bounds.y + bounds.height) - y1;
     set(x1, y1, w, h);
   }
